Move common functionality that is required both for PrivateAccessor and NonPublicExpectations into separate file as part of MockingUtil namespace.
