+++ Telerik.JustMock/Expectations/NonPublicExpectation.cs
@@ -27,6 +27,7 @@ using Telerik.JustMock.Core.Context;
 using Telerik.JustMock.Expectations.Abstraction;
 using Telerik.JustMock.Expectations.DynaMock;
 
+
 namespace Telerik.JustMock.Expectations
 {
 	internal sealed class NonPublicExpectation : INonPublicExpectation
@@ -119,12 +120,12 @@ namespace Telerik.JustMock.Expectations
 					}
 
 					if (mockedMethod == null)
-						throw new MissingMemberException(BuildMissingMethodMessage(type, mockedProperty, memberName));
+						throw new MissingMemberException(MockingUtil.BuildMissingMethodMessage(type, mockedProperty, memberName));
 				}
 			}
 
 			if (mockedMethod == null)
-				throw new MissingMemberException(BuildMissingMethodMessage(type, null, memberName));
+				throw new MissingMemberException(MockingUtil.BuildMissingMethodMessage(type, null, memberName));
 
 			if (mockedMethod.ContainsGenericParameters)
 			{
@@ -137,92 +138,6 @@ namespace Telerik.JustMock.Expectations
 			}
 
 			return mockedMethod;
-		}
-
-		private static string BuildMissingMethodMessage(Type type, PropertyInfo property, string memberName)
-		{
-			var sb = new StringBuilder();
-			sb.AppendLine(
-				property != null
-				? String.Format("Found property '{0}' on type '{1}' but the passed arguments match the signature neither of the getter nor the setter.", memberName, type)
-				: String.Format("Method '{0}' with the given signature was not found on type {1}", memberName, type));
-
-			var methods = new Dictionary<MethodInfo, string>();
-
-			if (property != null)
-			{
-				var getter = property.GetGetMethod(true);
-				if (getter != null)
-					methods.Add(getter, property.Name);
-				var setter = property.GetSetMethod(true);
-				if (setter != null)
-					methods.Add(setter, property.Name);
-			}
-
-			foreach (var method in type.GetMethods(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static).Where(m => m.Name == memberName))
-			{
-				methods.Add(method, method.Name);
-			}
-
-			if (methods.Count == 0)
-			{
-				sb.AppendLine("No methods or properties found with the given name.");
-			}
-			else
-			{
-				sb.AppendLine("Review the available methods in the message below and optionally paste the appropriate arrangement snippet.");
-				int i = 0;
-				foreach (var kvp in methods)
-				{
-					sb.AppendLine("----------");
-					var method = kvp.Key;
-					sb.AppendLine(String.Format("Method {0}: {1}", i + 1, method));
-					sb.AppendLine("C#: " + FormatMethodArrangementExpression(kvp.Value, method, SourceLanguage.CSharp));
-					sb.AppendLine("VB: " + FormatMethodArrangementExpression(kvp.Value, method, SourceLanguage.VisualBasic));
-					i++;
-				}
-			}
-
-			return sb.ToString();
-		}
-
-		private static string FormatMethodArrangementExpression(string memberName, MethodBase method, SourceLanguage language)
-		{
-			return String.Format("Mock.NonPublic.Arrange{0}({1}\"{2}\"{3}){4}",
-				FormatGenericArg(method.GetReturnType(), language),
-				method.IsStatic ? String.Empty : "mock, ",
-				memberName,
-				String.Join("", method.GetParameters().Select(p => ", " + FormatMethodParameterMatcher(p.ParameterType, language)).ToArray()),
-				language == SourceLanguage.CSharp ? ";" : "");
-		}
-
-		private static string FormatMethodParameterMatcher(Type paramType, SourceLanguage language)
-		{
-			if (paramType.IsByRef)
-			{
-				return String.Format("ArgExpr.Ref({0})", FormatMethodParameterMatcher(paramType.GetElementType(), language));
-			}
-			else
-			{
-				return String.Format("ArgExpr.IsAny{0}()", FormatGenericArg(paramType, language));
-			}
-		}
-
-		private static string FormatGenericArg(Type type, SourceLanguage language)
-		{
-			if (type == typeof(void))
-			{
-				return String.Empty;
-			}
-			switch (language)
-			{
-				case SourceLanguage.CSharp:
-					return String.Format("<{0}>", type.GetShortCSharpName());
-				case SourceLanguage.VisualBasic:
-					return String.Format("(Of {0})", type.GetShortVisualBasicName());
-				default:
-					throw new ArgumentOutOfRangeException("language");
-			}
 		}
 
 		private static MethodInfo FindMethodBySignature(IEnumerable<MethodInfo> candidates, Type returnType, object[] args)
@@ -249,50 +164,6 @@ namespace Telerik.JustMock.Expectations
 				});
 		}
 
-
-		private MethodInfo GetLocalMethod(Type type, MethodInfo method, string localMemberName)
-		{
-			MethodInfo[] allStaticNonPublicMethods = type.GetMethods(BindingFlags.Static | BindingFlags.NonPublic | BindingFlags.Instance);
-			MethodInfo[] potentialLocalMethods = allStaticNonPublicMethods.Where(m => (m.Name.Contains(method.Name) && m.Name.Contains(localMemberName))).ToArray();
-
-			if (potentialLocalMethods.Length == 0)
-			{
-				throw new MissingMemberException(BuildMissingMethodMessage(type, null, localMemberName));
-			} else if (potentialLocalMethods.Length == 1)
-			{
-				return potentialLocalMethods.First();
-			} else
-			{
-				MethodInfo localMethod = null;
-				MethodInfo[] allMethods = type.GetAllMethods().ToArray();
-				int methodIndex = Array.IndexOf(allMethods, method);
-				methodIndex += 1; //Indices in the names of local methods are 1-based
-
-				var regEx = new System.Text.RegularExpressions.Regex(@"<[a-z,A-z]>g__[a-z,A-z]|(?<MethodId>\d+)_(?<LocalId>\d+)");
-				foreach (var candidate in potentialLocalMethods)
-				{
-					System.Text.RegularExpressions.Match match = regEx.Match(candidate.Name);
-					if (match.Success)
-					{
-						int index = int.Parse(match.Groups["MethodId"].Value);
-						if (methodIndex == index)
-						{
-							localMethod = candidate;
-							break;
-						}
-					}
-				}
-				if (localMethod != null)
-				{
-					return localMethod;
-				}
-				else
-				{
-					throw new MissingMemberException(BuildMissingMethodMessage(type, null, localMemberName));
-				}
-			}
-		}
-
 		public ActionExpectation ArrangeLocal(object target, string memberName, string localMemberName, params object[] args)
 		{
 			Type[] emptyParamTypes = new Type[] {};
@@ -305,14 +176,14 @@ namespace Telerik.JustMock.Expectations
 			MethodInfo method = type.GetMethod(memberName, memberParamTypes);
 			if(method == null)
 			{
-				throw new MissingMemberException(BuildMissingMethodMessage(type, null, memberName));
+				throw new MissingMemberException(MockingUtil.BuildMissingMethodMessage(type, null, memberName));
 			}
 			return ArrangeLocal(target, method, localMemberName, args);
 		}
 		public ActionExpectation ArrangeLocal(object target, MethodInfo method, string localMemberName, params object[] args)
 		{
 			Type type = target.GetType();
-			MethodInfo localMethod = GetLocalMethod(type, method, localMemberName);
+			MethodInfo localMethod = MockingUtil.GetLocalMethod(type, method, localMemberName);
 
 			return Arrange(target, localMethod, args);
 		}
@@ -329,7 +200,7 @@ namespace Telerik.JustMock.Expectations
 			MethodInfo method = type.GetMethod(memberName, memberParamTypes);
 			if (method == null)
 			{
-				throw new MissingMemberException(BuildMissingMethodMessage(type, null, memberName));
+				throw new MissingMemberException(MockingUtil.BuildMissingMethodMessage(type, null, memberName));
 			}
 
 			return ArrangeLocal<TReturn>(target, method, localMemberName, args);
@@ -338,7 +209,7 @@ namespace Telerik.JustMock.Expectations
 		public FuncExpectation<TReturn> ArrangeLocal<TReturn>(object target, MethodInfo method, string localMemberName, params object[] args)
 		{
 			Type type = target.GetType();
-			MethodInfo localMethod = GetLocalMethod(type, method, localMemberName);
+			MethodInfo localMethod = MockingUtil.GetLocalMethod(type, method, localMemberName);
 
 			return Arrange<TReturn>(target, localMethod, args);
 		}
