+++ Telerik.JustMock/Telerik.JustMock.csproj
@@ -650,6 +650,7 @@
       <LastGenOutput>Args.Matching.cs</LastGenOutput>
     </None>
     <None Include="Core\DynamicProxy\Castle.DynamicProxy\DynProxy.snk" />
+    <Compile Include="Core\MockingUtil.LocalMethod.cs" />
     <None Include="Expectations\Abstraction\IReturns.tt">
       <Generator>TextTemplatingFileGenerator</Generator>
       <LastGenOutput>IReturns.cs</LastGenOutput>
